Move out common stores to separate ethStore
some parts of this code are now shared with the `Income` tab and are not exclusively used in the DAO anymore.
