+++ frontend/src/routes/Admin.svelte
@@ -1,18 +1,18 @@
 <script lang="ts">
-  import Navigation from "../components/Navigation.svelte";
-  import { API } from "../ts/api";
-  import Spinner from "../components/Spinner.svelte";
-  import { formatDate, formatNowUTC, storeToken } from "../ts/services";
-  import { config, error, loadedSponsoredRepos, user } from "../ts/mainStore";
   import {
-    faSignInAlt,
-    faCheck,
     faArrowsLeftRight,
+    faCheck,
+    faSignInAlt,
   } from "@fortawesome/free-solid-svg-icons";
   import Fa from "svelte-fa";
-  import type { Repo } from "../types/backend";
-  import Dots from "../components/Dots.svelte";
   import { navigate } from "svelte-routing";
+  import Dots from "../components/Dots.svelte";
+  import Navigation from "../components/Navigation.svelte";
+  import Spinner from "../components/Spinner.svelte";
+  import { API } from "../ts/api";
+  import { config, error, loadedSponsoredRepos, user } from "../ts/mainStore";
+  import { formatDate, formatNowUTC, storeToken } from "../ts/services";
+  import type { Repo } from "../types/backend";
 
   //let promisePendingPayouts =API.payouts.payoutInfos();
   let promiseTime = API.payouts.time();
