+++ frontend/src/routes/Income.svelte
@@ -7,7 +7,7 @@
   import { PayoutERC20ABI } from "../contracts/PayoutERC20";
   import { PayoutEthABI } from "../contracts/PayoutEth";
   import { API } from "../ts/api";
-  import { provider, signer } from "../ts/daoStore";
+  import { provider, signer } from "../ts/ethStore";
   import { error } from "../ts/mainStore";
   import { formatBalance, formatDate, timeSince } from "../ts/services";
   import type { PayoutResponse } from "../types/backend";
@@ -60,7 +60,7 @@
         ethSignature.s
       );
     } catch (exception) {
-      console.error(exception)
+      console.error(exception);
       $error = exception.data.data.reason;
       resetViewAfterPayout();
     }
